+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [PEP-440](https://www.python.org/dev/peps/pep-0440/)
 
 ## [Unreleased]
 
+### Added
+- Added dropdown filter for owner to blueprint list page
+
 ## [1.1.0] - 2021-01-15
 ### Changed
 - Significantly reduced load time of blueprints page for large datasets